changed variable names
