+++ index.js
@@ -40,7 +40,7 @@ app.use("/refresh", require("./routes/api/refresh"));
 app.use("/logout", require("./routes/api/logout"));
 app.use("/scrape", require("./routes/api/scrape"));
 app.use("/save", require("./routes/api/save")); // TODO: auth
-app.use("/param", require("./routes/api/param")) // TODO: users
+app.use("/pref", require("./routes/api/pref")); // TODO: users
 app.use(verifyJWT); // Everything below here requires user to be verified
 // app.use("/save", require("./routes/api/save")); // Save ads
