--- routes/api/param.js
@@ -1,10 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const paramController = require("../../controllers/paramController");
-
-router
-  .route("/")
-  .put(paramController.updateParam)
-  .get(paramController.getParam)
-
-  module.exports = router;
