+++ routes/api/pref.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const prefController = require("../../controllers/prefController");
+
+router.route("/").put(prefController.updatePref).get(prefController.getPref);
+
+module.exports = router;
